add unlink + refactor reducer + custom error class

+++ src/hooks/useAuthActions.jsx
@@ -1,6 +1,5 @@
 /** @format */
 import { ApiError, Okta } from '../common';
-import { actions } from '../providers/AuthProvider/AuthReducer';
 
 import {
 	getOAuthBaseUrl,
@@ -93,7 +92,7 @@ const useAuthActions = () => {
 			try {
 				const config = {};
 
-				dispatch({ type: actions.login.silentAuth.start.type });
+				dispatch({ type: 'SILENT_AUTH_STARTED' });
 
 				// Check if we can get tokens using either a refresh_token or, if our tokens are expired, if getWithoutPrompt works.
 				const isAuthenticated = await oktaAuth.isAuthenticated();
@@ -117,7 +116,7 @@ const useAuthActions = () => {
 							await oktaAuth.tokenManager.setTokens(tokens);
 
 							dispatch({
-								type: actions.login.silentAuth.success.type,
+								type: 'SILENT_AUTH_SUCCESS',
 								payload: { isAuthenticated },
 							});
 						}
@@ -125,14 +124,14 @@ const useAuthActions = () => {
 				}
 
 				dispatch({
-					type: actions.login.silentAuth.complete.type,
+					type: 'SILENT_AUTH_COMPLETED',
 				});
 
 				return { isAuthenticated };
 			} catch (error) {
 				if (dispatch) {
 					console.log(error);
-					dispatch({ type: actions.login.silentAuth.error.type, error });
+					dispatch({ type: 'SILENT_AUTH_ERROR', error });
 				} else {
 					throw new Error(error);
 				}
@@ -142,7 +141,7 @@ const useAuthActions = () => {
 		const getUser = async (dispatch, userId, _user) => {
 			try {
 				dispatch({
-					type: actions.user.fetch.start.type,
+					type: 'USER_FETCH_STARTED',
 					payload: { isLoadingUserProfile: true },
 				});
 
@@ -183,10 +182,15 @@ const useAuthActions = () => {
 					const _credentials = credentials.map(credential => {
 						const {
 							id,
-							provider: { id: idpId },
+							provider: { id: idpId, name },
 						} = credential;
 
-						return { ...credential, isLoggedIn: id === loggedInUserId && idpId === idp };
+						const isLoggedIn =
+							credentials.length === 1 ||
+							(credentials.length === 2 && ['email', 'password'].includes(name)) ||
+							(id === loggedInUserId && idpId === idp);
+
+						return { ...credential, isLoggedIn };
 					});
 
 					user = { ...user, credentials: _credentials };
@@ -194,7 +198,7 @@ const useAuthActions = () => {
 					localStorage.setItem('user', JSON.stringify(user));
 
 					dispatch({
-						type: actions.user.fetch.success.type,
+						type: 'USER_FETCH_SUCCEEDED',
 						payload: {
 							user: user,
 							isLoadingUserProfile: false,
@@ -208,7 +212,7 @@ const useAuthActions = () => {
 				if (dispatch) {
 					console.log(error);
 					dispatch({
-						type: actions.user.fetch.error.type,
+						type: 'USER_FETCH_FAILED',
 						error,
 					});
 				} else {
@@ -226,7 +230,7 @@ const useAuthActions = () => {
 				if (isAuthenticated) {
 					if (dispatch) {
 						dispatch({
-							type: actions.user.fetchInfo.start.type,
+							type: 'USER_INFO_FETCH_STARTED',
 							payload,
 						});
 					}
@@ -248,13 +252,13 @@ const useAuthActions = () => {
 					}
 
 					if (dispatch) {
-						dispatch({ type: actions.user.fetchInfo.success.type, payload });
+						dispatch({ type: 'USER_INFO_FETCH_SUCCEEDED', payload });
 					}
 				}
 			} catch (error) {
 				if (dispatch) {
 					console.log(error);
-					dispatch({ type: actions.user.fetchInfo.error.type, error });
+					dispatch({ type: 'USER_INFO_FETCH_FAILED', error });
 				} else {
 					throw new Error(error);
 				}
@@ -265,15 +269,15 @@ const useAuthActions = () => {
 			try {
 				console.debug('generating URL...');
 
-				dispatch({ type: actions.login.start.withModal.type });
+				dispatch({ type: 'LOGIN_START_WITH_MODAL' });
 
 				const { authUrl, tokenParams } = await generateAuthUrl(oktaAuth, options);
 
-				return dispatch({ type: actions.login.pending.type, payload: { authUrl, tokenParams } });
+				return dispatch({ type: 'LOGIN_PENDING', payload: { authUrl, tokenParams } });
 			} catch (error) {
 				if (dispatch) {
 					console.log('loginWithModal error:', error);
-					dispatch({ type: actions.login.error.type, error });
+					dispatch({ type: 'LOGIN_ERROR', error });
 				} else {
 					throw new Error(error);
 				}
@@ -282,7 +286,7 @@ const useAuthActions = () => {
 
 		const loginWithCredentials = async (dispatch, { username, password }) => {
 			try {
-				dispatch({ type: actions.login.start.withCredentials.type });
+				dispatch({ type: 'LOGIN_WITH_CREDENTIALS_STARTED' });
 
 				oktaAuth
 					.signInWithCredentials({ username, password, sendFingerprint: true })
@@ -293,7 +297,7 @@ const useAuthActions = () => {
 					});
 			} catch (error) {
 				if (dispatch) {
-					dispatch({ type: actions.login.error.type, error });
+					dispatch({ type: 'LOGIN_ERROR', error });
 				} else {
 					throw new Error(error);
 				}
@@ -311,7 +315,7 @@ const useAuthActions = () => {
 				}
 
 				if (idp) {
-					dispatch({ type: actions.login.start.withRedirect.type });
+					dispatch({ type: 'LOGIN_WITH_REDIRECT_STARTED' });
 
 					return await oktaAuth.signInWithRedirect({ idp: idpMap[idp] });
 				}
@@ -324,7 +328,7 @@ const useAuthActions = () => {
 
 					if (!response?.tokens) {
 						return dispatch({
-							type: actions.login.error.type,
+							type: 'LOGIN_ERROR',
 							error: `No tokens in response. Something went wrong! [${response}]`,
 						});
 					}
@@ -333,11 +337,11 @@ const useAuthActions = () => {
 
 					await oktaAuth.authStateManager.updateAuthState();
 
-					return dispatch({ type: actions.login.success.type, payload: { isAuthenticated: true } });
+					return dispatch({ type: 'LOGIN_SUCCESS', payload: { isAuthenticated: true } });
 				}
 
 				if (oktaAuth.isLoginRedirect() || tokens) {
-					dispatch({ type: actions.login.pending.type });
+					dispatch({ type: 'LOGIN_PENDING' });
 
 					await oktaAuth.handleLoginRedirect();
 
@@ -347,7 +351,7 @@ const useAuthActions = () => {
 
 					// await oktaAuth.authStateManager.updateAuthState();
 
-					dispatch({ type: actions.login.success.type });
+					dispatch({ type: 'LOGIN_SUCCESS' });
 
 					return await getUserInfo(dispatch);
 				}
@@ -367,7 +371,7 @@ const useAuthActions = () => {
 				}
 			} catch (error) {
 				if (dispatch) {
-					dispatch({ type: actions.login.error.type, error });
+					dispatch({ type: 'LOGIN_ERROR', error });
 				} else {
 					throw new Error(error);
 				}
@@ -396,31 +400,32 @@ const useAuthActions = () => {
 			if (postLogoutRedirect) {
 				config = { postLogoutRedirectUri: postLogoutRedirect };
 			}
+			dispatch({ type: 'LOGOUT_STARTED' });
 
 			console.info('executing logout...');
 
 			localStorage.removeItem('user');
 
-			return oktaAuth.signOut(config).then(() => dispatch({ type: 'LOGOUT_SUCCESS' }));
+			return oktaAuth.signOut(config).then(() => dispatch({ type: 'LOGOUT_SUCCEEDED' }));
 		};
 
 		const toggleSignUp = (dispatch, isSignUp) =>
 			dispatch({
-				type: actions.login.init.signUp.type,
+				type: 'LOGIN_INIT_SIGN_UP'.type,
 				payload: { isSignUp: !isSignUp },
 			});
 
 		const toggleEmailAuth = (dispatch, { isRecovery, isSignUp, isEmailAuth }) => {
 			if (isRecovery) {
 				dispatch({
-					type: actions.login.init.withEmail.type,
+					type: 'LOGIN_INIT_WITH_EMAIL'.type,
 					payload: { isRecovery: !isRecovery, isEmailAuth: false },
 				});
 			} else if (isSignUp) {
-				dispatch({ type: actions.login.init.signUp.type, payload: { isEmailAuth: !isEmailAuth } });
+				dispatch({ type: 'LOGIN_INIT_SIGN_UP'.type, payload: { isEmailAuth: !isEmailAuth } });
 			} else {
 				dispatch({
-					type: actions.login.init.withEmail.type,
+					type: 'LOGIN_INIT_WITH_EMAIL'.type,
 					payload: { isEmailAuth: !isEmailAuth },
 				});
 			}
@@ -436,10 +441,7 @@ const useAuthActions = () => {
 				if (oktaAuth.isLoginRedirect()) {
 					// const { state } = await oktaAuth.parseOAuthResponseFromUrl(oktaAuth);
 					dispatch({
-						type: actions.user.link.pending.type,
-						payload: {
-							isLoadingLinkProfile: true,
-						},
+						type: 'USER_LINK_PENDING',
 					});
 
 					const primaryAccessToken = oktaAuth.getAccessToken();
@@ -449,14 +451,9 @@ const useAuthActions = () => {
 						payload: { uid },
 					} = await oktaAuth.token.decode(primaryAccessToken);
 
-					console.log('=== primaryAccessToken ===');
-					console.log(primaryAccessToken);
-
 					await oktaAuth.handleLoginRedirect();
 
 					const linkAccessToken = oktaAuth.getAccessToken();
-					console.log('=== linkAccessToken ===');
-					console.log(linkAccessToken);
 
 					const options = {
 						method: 'post',
@@ -479,17 +476,11 @@ const useAuthActions = () => {
 					await oktaAuth.tokenManager.renew('accessToken');
 
 					dispatch({
-						type: actions.user.link.success.type,
-						payload: {
-							isLoadingLinkProfile: false,
-							isStaleUserProfile: true,
-						},
+						type: 'USER_LINK_SUCCEEDED',
 					});
 
 					// restore the redirect_uri
 					oktaAuth.options.redirectUri = `${window.location.origin}/login/callback`;
-					console.log('=== CURRENT REDIRECT_URI ===');
-					console.log(oktaAuth.options.redirectUri);
 
 					return;
 				}
@@ -501,19 +492,24 @@ const useAuthActions = () => {
 				// 	});
 
 				// 	return dispatch({
-				// 		type: actions.user.link.start.type,
+				// 		type: 'USER_LINK_STARTED',
 				// 		payload: { authUrl, tokenParams },
 				// 	});
 				else {
 					const scopes = oktaAuth.options.scopes;
 
 					dispatch({
-						type: actions.user.link.start.type,
-						payload: { isLoadingLinkProfile: true },
+						type: 'USER_LINK_STARTED',
 					});
+
+					console.info(' ===== CURRENT ORIGINAL URI ===== ');
+					console.log(oktaAuth.getOriginalUri());
 
 					// Ensures we reroute to the /settings page where we started.
 					await oktaAuth.setOriginalUri(window.location.href);
+
+					console.info(' ===== NOW RETURNING TO ===== ');
+					console.log(oktaAuth.getOriginalUri());
 
 					const options = {
 						idp: idpMap[idp],
@@ -534,18 +530,31 @@ const useAuthActions = () => {
 				// call /api/v1/users/{{userId}}/identities
 			} catch (error) {
 				if (dispatch) {
-					dispatch({ type: actions.user.link.error.type, error });
+					dispatch({ type: 'USER_LINK_FAILED', error });
 				} else {
 					throw new Error(error);
 				}
 			}
 		};
 
-		const unlinkUser = async (dispatch, primaryId, associatedId) => {
+		const unlinkUser = async (dispatch, credential) => {
 			try {
-				dispatch({ type: actions.user.unlink.start.type });
+				dispatch({ type: 'USER_UNLINK_STARTED' });
 
-				const url = `${window.location.origin}/api/v1/users/${associatedId}identities`;
+				const {
+					id,
+					provider: { id: idpId },
+					isLoggedIn,
+				} = credential;
+
+				if (isLoggedIn) {
+					throw new Error('Cannot disconnect the currently logged in account.');
+				}
+
+				const baseUrl = `${window.location.origin}/api/v1/users/${id}/identities`;
+
+				const url = idpId ? baseUrl + `/${idpId}` : baseUrl;
+
 				const options = {
 					method: 'delete',
 					headers: {
@@ -555,18 +564,18 @@ const useAuthActions = () => {
 
 				const response = await fetch(url, options);
 
-				if (response.statusCode !== 204) {
+				if (response.status !== 204) {
 					throw new ApiError({
 						statusCode: response.statusCode,
-						message: `Unable to unlink user ${associatedId} from ${primaryId}`,
-						json: await response.json(),
+						message: `Unable to unlink user ${id}`,
+						json: (await response.json()) || '',
 					});
 				}
 
-				return dispatch({ type: actions.user.unlink.success.type });
+				return dispatch({ type: 'USER_UNLINK_SUCCEEDED' });
 			} catch (error) {
 				if (dispatch) {
-					dispatch({ type: actions.user.unlink.error.type, error });
+					dispatch({ type: 'USER_UNLINK_FAILED', error });
 				} else {
 					throw new Error(error);
 				}
